// src/selectors/kpi/kpi-data/models/IKpiDataDetailPointTableRow.ts
export default interface IKpiDataDetailPointTableRow {
  readonly docno: string;
  readonly emplid: number;
  readonly udcid: number;
  readonly doctype: string;
  readonly creator: string;
  readonly remark: string;
  readonly bu: string;
  readonly customer: string;
  readonly startDate: string;
  readonly endDate: string;
  readonly point: number;
  readonly rnum2: number;
}